Use xmltodict instead of ElementTree
ElementTree is awkward to navigate, due to having to dig into the tag
variable of each item in order to discover the blocks' tag. Navigating
the XML as a dictionary is much more natural and readable.

I've also refactored the code to have a more logical separation, which
should make it easier to read/maintain in the future.